# Remove unused "nodes_from_edges" computation