--- a/stellargraph/core/graph.py
+++ b/stellargraph/core/graph.py
@@ -319,10 +319,6 @@ class StellarGraph:
             nodes=self._nodes,
         )
 
-        nodes_from_edges = pd.unique(
-            np.concatenate([self._edges.targets, self._edges.sources])
-        )
-
     @staticmethod
     def _infer_nodes_from_edges(edges, source_column, target_column):
         # `convert_edges` nicely flags any errors in edges; inference here is lax rather than duplicate that